Reduce code duplication between v3 and v4 bundle
